Eliminate semantic usize at serialization boundaries (closes #12)
All serde-serialized struct fields in uor-r4-router converted usize to
u64 (result structs, router state maps, facet store, sparse wire repr);
UorAttestationResult.hash_algorithm_id likewise. JSON wire shape
identical (numbers), proven by new exact-JSON round-trip tests;
pre-change manifold_cache_rust.json decodes fine. get_expert_counts
keeps its usize wasm ABI deliberately (Vec<u64> would switch JS to
BigInt elements). Verified: 43 suites, clippy -D warnings, fmt.

+++ crates/uor-r4-core/src/lib.rs
@@ -57,7 +57,7 @@ impl R4Vector {
 pub struct UorAttestationResult {
     pub algorithm: String,
     pub hash_algorithm: String,
-    pub hash_algorithm_id: usize,
+    pub hash_algorithm_id: u64,
     pub address: String,
     pub kappa_label: String,
     pub fingerprint_hex: String,
